requirements.txt
supabase>=0.2.2
python-dotenv>=0.19.2
numpy
scipy